--- a/be/src/exec/vectorized/hdfs_scanner_orc.cpp
+++ b/be/src/exec/vectorized/hdfs_scanner_orc.cpp
@@ -15,7 +15,7 @@ class ORCHdfsFileStream : public orc::InputStream {
 public:
     // |file| must outlive ORCHdfsFileStream
     ORCHdfsFileStream(RandomAccessFile* file, uint64_t length, HdfsScanStats* stats)
-            : _file(std::move(file)), _length(length), _stats(stats) {}
+            : _file(std::move(file)), _length(length), _stats(stats), _small_buffer(0), _small_buffer_offset(0) {}
 
     ~ORCHdfsFileStream() override = default;
 
@@ -40,7 +40,32 @@ public:
 
     uint64_t getNaturalReadSizeAfterSeek() const override { return 256 * 1024; }
 
+    void prepareCache(orc::InputStream::PrepareCacheEvent event, uint64_t offset, uint64_t length) override {
+        if (length > kSmallBufferMaxSize) return;
+        if (can_use_small_buffer(offset, length)) return;
+        _small_buffer.resize(length);
+        _small_buffer_offset = offset;
+        doRead(_small_buffer.data(), length, offset);
+    }
+
+    inline bool can_use_small_buffer(uint64_t offset, uint64_t length) {
+        if ((_small_buffer.size() != 0) && (offset >= _small_buffer_offset) &&
+            ((offset + length) <= (_small_buffer_offset + _small_buffer.size()))) {
+            return true;
+        }
+        return false;
+    }
+
     void read(void* buf, uint64_t length, uint64_t offset) override {
+        if (can_use_small_buffer(offset, length)) {
+            size_t idx = offset - _small_buffer_offset;
+            memcpy(buf, _small_buffer.data() + idx, length);
+        } else {
+            doRead(buf, length, offset);
+        }
+    }
+
+    void doRead(void* buf, uint64_t length, uint64_t offset) {
         SCOPED_RAW_TIMER(&_stats->io_ns);
         _stats->io_count += 1;
         if (buf == nullptr) {
@@ -58,9 +83,12 @@ public:
     const std::string& getName() const override { return _file->filename(); }
 
 private:
+    static const uint64_t kSmallBufferMaxSize = 2 * 1024 * 1024;
     RandomAccessFile* _file;
     uint64_t _length;
     HdfsScanStats* _stats;
+    std::vector<char> _small_buffer;
+    uint64_t _small_buffer_offset;
 };
 
 class OrcRowReaderFilter : public orc::RowReaderFilter {