@DorianZheng 2022-03-09
Is it better to user `BufferedInputStream::BufferedInputStream` directly