@dirtysalt 2022-03-09
> Is it better to use `BufferedInputStream::BufferedInputStream` for buffering `RandomAccessFile` directly

That works. I'll look into this class.