@dirtysalt 2022-03-09
> Is it better to use `BufferedInputStream::BufferedInputStream` for buffering `RandomAccessFile` directly

I've checked `BufferedInputStream` implementation. I think it's cumbersome to use BufferedInputStream here, and I don't it makes thing simpler.w